+++ src/js/components/nestable.js
@@ -5,13 +5,13 @@
 
     var component;
 
-    if (window.UIkit) {
-        component = addon(UIkit);
+    if (window.UIkit2) {
+        component = addon(UIkit2);
     }
 
     if (typeof define == 'function' && define.amd) {
         define('uikit-nestable', ['uikit'], function(){
-            return component || addon(UIkit);
+            return component || addon(UIkit2);
         });
     }