+++ src/js/core/nav.js
@@ -149,4 +149,4 @@
         return height;
     }
 
-})(UIkit);
+})(UIkit2);