+++ src/js/core/tab.js
@@ -90,7 +90,7 @@
 
             // init UIkit components
             if (this.options.connect) {
-                
+
                 this.switcher = UI.switcher(this.element, {
                     toggle    : '>li:not(.uk-tab-responsive)',
                     connect   : this.options.connect,
@@ -165,4 +165,4 @@
         }
     });
 
-})(UIkit);
+})(UIkit2);