+++ src/js/core/utility.js
@@ -331,4 +331,4 @@
         return val;
     }
 
-})(UIkit);
+})(UIkit2);