+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Changelog
 
+
+### WIP
+
+  - Added global object UIkit2, so internal code is not dependent on the UIkit object
+  - Updated prefix behaviour to also create a global UIkit2{prefix} object
+
 ### 2.27.2 (Oct 24, 2016)
 
   - Updated datepicker default value when minDate setting is set
-  - Updated tooltip word-break behavior 
+  - Updated tooltip word-break behavior
   - Fixed slideshow bug when using slideshow-fx based effects
   - Fixed dropdown error if now offset element exists
   - Fixed dropdowns on windows with touch support